refactor: 파일 이동으로 생긴 오류 해결(#289)
