+++ src/main/java/com/soda/global/init/InitialDataLoader.java
@@ -2,8 +2,8 @@ package com.soda.global.init;
 
 import com.soda.member.domain.Member;
 import com.soda.member.domain.MemberRole;
-import com.soda.member.repository.CompanyRepository;
-import com.soda.member.repository.MemberRepository;
+import com.soda.member.infrastructure.CompanyRepository;
+import com.soda.member.infrastructure.MemberRepository;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
