+++ src/main/java/com/soda/member/domain/AuthService.java
@@ -8,8 +8,8 @@ import com.soda.member.interfaces.dto.VerificationConfirmResponse;
 import com.soda.member.interfaces.dto.member.LoginRequest;
 import com.soda.member.interfaces.dto.member.LoginResponse;
 import com.soda.member.interfaces.dto.member.admin.CreateMemberRequest;
-import com.soda.member.repository.RefreshTokenRepository;
-import com.soda.member.repository.VerificationCodeRepository;
+import com.soda.member.infrastructure.RefreshTokenRepository;
+import com.soda.member.infrastructure.VerificationCodeRepository;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.MalformedJwtException;
 import io.jsonwebtoken.UnsupportedJwtException;
