+++ src/main/java/com/soda/member/infrastructure/member/MemberProviderImpl.java
@@ -2,7 +2,7 @@ package com.soda.member.infrastructure.member;
 
 import com.soda.member.domain.member.MemberProvider;
 import com.soda.member.domain.Member;
-import com.soda.member.repository.MemberRepository;
+import com.soda.member.infrastructure.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
